refactor & tidy
