+++ packages/common/src/streams.ts
@@ -68,6 +68,12 @@ export const bytesToStream = (bytes: Uint8Array): Readable => {
   return stream
 }
 
+export async function* bytesToIterable(
+  bytes: Uint8Array,
+): AsyncIterable<Uint8Array> {
+  yield bytes
+}
+
 export class MaxSizeChecker extends Transform {
   totalSize = 0
   constructor(
