+++ packages/pds/src/api/com/atproto/sync/getRepo.ts
@@ -1,4 +1,5 @@
 import stream from 'node:stream'
+import { byteIterableToStream } from '@atproto/common'
 import { InvalidRequestError } from '@atproto/xrpc-server'
 import {
   RepoRootNotFoundError,
@@ -41,7 +42,8 @@ export const getCarStream = async (
   let carStream: stream.Readable
   try {
     const storage = new SqlRepoReader(actorDb)
-    carStream = await storage.getCarStream(since)
+    const carIter = await storage.getCarStream(since)
+    carStream = byteIterableToStream(carIter)
   } catch (err) {
     await actorDb.close()
     if (err instanceof RepoRootNotFoundError) {
