+++ packages/repo/src/index.ts
@@ -7,4 +7,5 @@ export * from './storage'
 export * from './sync'
 export * from './types'
 export * from './data-diff'
+export * from './car'
 export * from './util'
